refactor(viewer): split StorageBrowser into components/storage/browser
3,001-line panel becomes a 414-line composition over ./browser: row
components (FileRow, FolderRow, ProceduralModelRow, VersionsTree), the
header, selection toolbar, upload/progress strip, FileTreeList and the
dialogs; hooks for folder state, uploads, procedural models, file ops,
bulk actions, drag & drop + menu builders, and keyboard navigation; and
the versions/key helpers. Markup, class names and strings unchanged.
No async poll loop lived here (only the refresh spin window and the
long-press timer), so nothing to move onto useJobPoll.
